+++ app/src/main/java/ru/gdcn/beastmaster64revelations/GameActivities/CharacterCreation/AttributeCounter.java
@@ -1,4 +1,4 @@
-package ru.gdcn.beastmaster64revelations.CharacterCreation;
+package ru.gdcn.beastmaster64revelations.GameActivities.CharacterCreation;
 
 import android.widget.TextView;
