+++ app/src/main/java/ru/gdcn/beastmaster64revelations/GameActivities/CharacterCreation/CharacterCreationActivity.java
@@ -1,24 +1,16 @@
-package ru.gdcn.beastmaster64revelations.CharacterCreation;
+package ru.gdcn.beastmaster64revelations.GameActivities.CharacterCreation;
 
 import android.content.Intent;
-import android.media.Image;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.text.Editable;
 import android.text.TextWatcher;
-import android.view.View;
-import android.view.Window;
-import android.view.WindowManager;
 import android.widget.Button;
-import android.widget.FrameLayout;
 import android.widget.ImageView;
 import android.widget.TextView;
 
 import ru.gdcn.beastmaster64revelations.GameActivities.InLocation.InLocationActivity;
 import ru.gdcn.beastmaster64revelations.GameClass.Characters.PlayerClass;
-import ru.gdcn.beastmaster64revelations.GameInterface.Character.Character;
 import ru.gdcn.beastmaster64revelations.GameInterface.Character.NPC.Player;
 import ru.gdcn.beastmaster64revelations.R;
