+++ app/src/main/java/ru/gdcn/beastmaster64revelations/GameActivities/CharacterCreation/CharacterCreationFragment.java
@@ -1,4 +1,4 @@
-package ru.gdcn.beastmaster64revelations.CharacterCreation;
+package ru.gdcn.beastmaster64revelations.GameActivities.CharacterCreation;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -11,8 +11,6 @@ import android.view.ViewGroup;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
-import ru.gdcn.beastmaster64revelations.GameClass.Characters.PlayerClass;
-import ru.gdcn.beastmaster64revelations.GameInterface.Character.NPC.Player;
 import ru.gdcn.beastmaster64revelations.R;
 
 public class CharacterCreationFragment extends Fragment {
