+++ app/src/main/java/ru/gdcn/beastmaster64revelations/GameActivities/CharacterCreation/Gender.java
@@ -1,4 +1,4 @@
-package ru.gdcn.beastmaster64revelations.CharacterCreation;
+package ru.gdcn.beastmaster64revelations.GameActivities.CharacterCreation;
 
 import java.io.Serializable;
