+++ app/src/main/java/ru/gdcn/beastmaster64revelations/GameActivities/InLocation/InLocationActivity.java
@@ -11,20 +11,21 @@ import android.support.v7.app.AppCompatActivity;
 import android.widget.Button;
 import android.widget.FrameLayout;
 
-import ru.gdcn.beastmaster64revelations.FightActivity;
+import ru.gdcn.beastmaster64revelations.GameActivities.FightActivity;
 import ru.gdcn.beastmaster64revelations.GameClass.Characters.PlayerClass;
 import ru.gdcn.beastmaster64revelations.GameClass.WorldElemets.SimpleLocationClass;
 import ru.gdcn.beastmaster64revelations.GameClass.WorldElemets.SimpleWorldClass;
+import ru.gdcn.beastmaster64revelations.GameFragments.InLocationFragment;
 import ru.gdcn.beastmaster64revelations.GameInterface.Character.Interactions.Fight.FightResult;
 import ru.gdcn.beastmaster64revelations.GameInterface.Character.NPC.Player;
 import ru.gdcn.beastmaster64revelations.GameInterface.World.Location.Location;
 import ru.gdcn.beastmaster64revelations.GameInterface.World.Location.LocationType;
 import ru.gdcn.beastmaster64revelations.GameInterface.World.MapDirection;
 import ru.gdcn.beastmaster64revelations.GameInterface.World.World;
-import ru.gdcn.beastmaster64revelations.MainMenu.MainMenuActivity;
-import ru.gdcn.beastmaster64revelations.MapFragment;
+import ru.gdcn.beastmaster64revelations.GameActivities.MainMenu.MainMenuActivity;
+import ru.gdcn.beastmaster64revelations.GameFragments.MapFragment.MapFragment;
 import ru.gdcn.beastmaster64revelations.R;
-import ru.gdcn.beastmaster64revelations.StatsFragment;
+import ru.gdcn.beastmaster64revelations.GameFragments.StatsFragment;
 import ru.gdcn.beastmaster64revelations.UIElements.ProportionalImageView;
 
 public class InLocationActivity extends AppCompatActivity {
