+++ app/src/main/res/layout/activity_fight.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#333"
-    tools:context=".FightActivity">
+    tools:context=".GameActivities.FightActivity">
 
     <!-- The primary full-screen view. This can be replaced with whatever view
          is needed to present your content, e.g. VideoView, SurfaceView,
