+++ app/src/main/res/layout/fragment_character_creation.xml
@@ -9,7 +9,7 @@
     android:gravity="top|center"
     android:orientation="vertical"
     android:showDividers="beginning|middle|end"
-    tools:context=".CharacterCreation.CharacterCreationFragment">
+    tools:context=".GameActivities.CharacterCreation.CharacterCreationFragment">
 
 
     <TextView
