+++ app/src/main/res/layout/fragment_in_location.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/my_border"
 
-    tools:context=".GameActivities.InLocation.InLocationFragment">
+    tools:context=".GameFragments.InLocationFragment">
 
     <ScrollView
         android:id="@+id/fragment_in_location_scrollContent"
