+++ app/src/main/res/xml/pref_headers.xml
@@ -3,17 +3,17 @@
     <!-- These settings headers are only used on tablets. -->
 
     <header
-        android:fragment="ru.gdcn.beastmaster64revelations.Settings.SettingsActivity$GeneralPreferenceFragment"
+        android:fragment="ru.gdcn.beastmaster64revelations.GameActivities.Settings.SettingsActivity$GeneralPreferenceFragment"
         android:icon="@drawable/ic_info_black_24dp"
         android:title="@string/pref_header_general" />
 
     <header
-        android:fragment="ru.gdcn.beastmaster64revelations.Settings.SettingsActivity$NotificationPreferenceFragment"
+        android:fragment="ru.gdcn.beastmaster64revelations.GameActivities.Settings.SettingsActivity$NotificationPreferenceFragment"
         android:icon="@drawable/ic_notifications_black_24dp"
         android:title="@string/pref_header_notifications" />
 
     <header
-        android:fragment="ru.gdcn.beastmaster64revelations.Settings.SettingsActivity$DataSyncPreferenceFragment"
+        android:fragment="ru.gdcn.beastmaster64revelations.GameActivities.Settings.SettingsActivity$DataSyncPreferenceFragment"
         android:icon="@drawable/ic_sync_black_24dp"
         android:title="@string/pref_header_data_sync" />
